requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
actions-python-core==0.1.1
    # via
    #   -r requirements.in
    #   actions-python-github
actions-python-github==0.1.1
    # via -r requirements.in
aiofiles==23.2.1
    # via actions-python-core
annotated-types==0.5.0
    # via pydantic
anyio==4.0.0
    # via httpcore
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
exceptiongroup==1.1.3
    # via anyio
githubkit==0.11.0a0
    # via actions-python-github
h11==0.14.0
    # via httpcore
httpcore==0.18.0
    # via httpx
httpx==0.25.0
    # via
    #   actions-python-core
    #   githubkit
idna==3.4
    # via
    #   anyio
    #   httpx
pydantic==2.4.2
    # via
    #   actions-python-github
    #   githubkit
pydantic-core==2.10.1
    # via pydantic
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
typing-extensions==4.8.0
    # via
    #   actions-python-core
    #   githubkit
    #   pydantic
    #   pydantic-core
